Move legacy definition code (WIP)
