+++ temba/flows/legacy/__init__.py
@@ -1,1 +1,2 @@
+from .definition import *  # noqa
 from .engine import *  # noqa
